issue-24: moviendo formulario a otro componente para hacer el cambio mas simple
